Update Vote.php
from ussignBigInteger to bigInteger

+++ app/Models/Vote.php
@@ -15,8 +15,8 @@ class Vote extends Model
     
 
     protected $casts = [
-        'id' => 'unsignedBigInteger' , 
-        'user_id' => 'integer' , 
+        'id' => 'BigInteger' , 
+        'user_id' => 'BigInteger' , 
         'vote' => 'integer' , 
     ];
